--- MDFactory/Sources/Scenes/MDF-4/View/Enums/Metric.swift
@@ -1,17 +1,0 @@
-//
-//  Metriks.swift
-//  MDFactory
-//
-//  Created by Andrei Maskal on 12/09/2022.
-//
-
-import UIKit
-
-
-enum Metric {
-    static var playerHeight: CGFloat = 90
-    static var screenHeight = UIScreen.main.bounds.height
-
-}
-
-
